requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
defusedxml==0.5.0
future==0.16.0
iso-639==0.4.5
iso3166==0.8
lxml==3.7.3
numpy==1.12.1
oauthlib==2.0.2
olefile==0.44
opencv-contrib-python==3.2.0.7
Pillow==4.1.0
prawcore==0.8.0
pycryptodome==3.4.5
pytesseract==0.1.6
python-twitter==3.2.1
requests==2.13.0
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
retrying==1.3.3
scipy==0.19.0
six==1.10.0
streamlink==0.5.0
update-checker==0.16
